requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
chardet==5.2.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-tailwind==0.5.0
Django==4.2.16
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.3
django-mssql-backend==2.8.1
djangorestframework==3.15.2
et_xmlfile==2.0.0
iniconfig==2.1.0
kombu==5.5.4
mssql-django==1.5
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==10.4.0
pluggy==1.6.0
prompt_toolkit==3.0.51
Pygments==2.19.2
pyodbc==5.2.0
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
redis==5.2.0
reportlab==4.2.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
holidays==0.58
